Move app callback to Mongo.App
